[part 8] refactor!(telemetry): rework metrics API (#206)
Redesign the metrics API to eliminate the need for pre-registering a list
of metrics.

Previously, the API relied on a pull-based approach where metrics had to be
registered during the construction of the telemetry module. The module would
then periodically retrieve values from the registered metrics.
While functional, this design was cumbersome—especially for metrics involving
dynamic tags.

This refactoring introduces a push-based mechanism. Calls to `rate`, `counter`,
or `distribution` now directly send data to the telemetry module, removing
the need for upfront registration.
This approach simplifies usage and improves flexibility.

* simplify test and send metrics in app-closing

* add counter test

* fix typo

* fix json equality

* fix json ordering 2

* fix windows and formating

* fix windows 2

* Apply suggestions from code review

Co-authored-by: pablomartinezbernardo <134320516+pablomartinezbernardo@users.noreply.github.com>

---------

Co-authored-by: pablomartinezbernardo <134320516+pablomartinezbernardo@users.noreply.github.com>