fix(rc): rework remote configuration handling (#234)
This commit refactors the way remote configurations are applied
to the tracer. Previously, if a remote configuration contained both
valid and invalid fields, the valid parts could be partially applied,
potentially leading to inconsistent behavior.

Now, the configuration is only applied if it is fully valid. Any invalid
configuration is entirely rejected. This constitutes a breaking change.

This also report errors when an invalid configuration is rejected.

Changes:
  - Add fuzz test and input corpus for Remote Configuration.
  - Fix a crash when telemetry methods are invoked before intializing
    the module.
  - Improve unit test coverage of remote configuration.