feat: Allow the "manual.keep" tag to override the tracer's sampling decision.
If the user sets the span tag "manual.keep" with a truthy value, this overrides the sampling priority of the span to USER_KEEP rather than setting a tag value.

This brings dd-trace-cpp in line with other Datadog tracers to allow users the flexibility to override a sampling decision using the span.set_tag() API.

Note: This change is breaking in the sense that a user can no longer set a "manual.keep" tag as it is no longer saved to the span's meta tags.